pyproject.toml
[project]
name = "task-manager-bels"
version = "0.1.0"
description = ""
authors = [
    {name = "KirillBels",email = "k.belskiy@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
